break tunneler callbacks out of application module
