+++ programs/ziti-tunneler/CMakeLists.txt
@@ -1,6 +1,5 @@
 project(ziti-tunneler-app)
 
-
 if(CMAKE_SYSTEM_NAME STREQUAL Darwin)
     set(NETIF_DRIVER_SOURCE netif_driver/darwin/utun.c)
 endif()
@@ -8,7 +7,7 @@ if(CMAKE_SYSTEM_NAME STREQUAL Linux)
     set(NETIF_DRIVER_SOURCE netif_driver/linux/tun.c)
 endif()
 
-add_executable(ziti-tunneler-app ziti-tunneler.c ${NETIF_DRIVER_SOURCE})
+add_executable(ziti-tunneler-app ziti-tunneler.c ziti-tunneler-cbs.c ${NETIF_DRIVER_SOURCE})
 
 target_link_libraries(ziti-tunneler-app
         PUBLIC ziti ziti_tunneler
